Move highlight method into utilities/helpers.py
