+++ bi_ui/utilities/helpers.py
@@ -1,4 +1,5 @@
 from functools import reduce
+from flask import Markup
 
 
 # https://mathieularose.com/function-composition-in-python/
@@ -11,3 +12,14 @@ def convert_band(business: dict, key: str, not_found_key: str, bands: dict) -> d
     not_found_msg = f'No {not_found_key} could be found.'
     description = bands.get(initial_value, not_found_msg)
     return {**business, key: f'{initial_value} - {description}'}
+
+
+def highlight(business: dict, to_highlight: str) -> dict:
+    original = business['BusinessName']
+    # We need to use Markup so that Jinja will render our HTML properly
+    new_name = Markup(original.replace(to_highlight.upper(), f'<em class="highlight">{to_highlight.upper()}</em>'))
+    highlighted_business = {
+        **business,
+        "BusinessName": new_name
+    }
+    return highlighted_business
